1.添加图像消息
git-svn-id: http://10.22.1.27:8000/svn/rooten/MSMS/src/client/IOS/IM@940 47c12b7e-2652-5545-b168-31c46e2fce3f

+++ IM/fileMgr/FileBlocksMgr.m
@@ -71,7 +71,7 @@ static NSString *kBlocksSuffix = @".blocksInfo";
 }
 
 - (BOOL)creatBlocksAtPath:(NSString *)path {
-    NSUInteger blocksCount = m_sz / m_defaultBlockSz;
+    unsigned long long blocksCount = m_sz / m_defaultBlockSz;
     NSUInteger left = m_sz % m_defaultBlockSz;
     NSUInteger n = 0;
     for (; n < blocksCount; n++) {
